toxic_classifier/requirements.txt
django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.0
torch==2.1.0
transformers==4.36.0
peft==0.7.0
numpy==1.24.3
tqdm==4.66.1